ASoC: clearwater: Apply constraints for "* Rate" controls
We must not modify the rate feeding a particular mixer if any of the inputs
to that mixer are non-zero (non-mute) There is also a strict timing constraint
for modifying the the rate of a mixer block where any of the mixer sources
were "recently" non-zero, "recently" is defined as 1 sample period. Our worst
case sample rate is 8kHz which results in a sample period of 125 uS.

In order to avoid the rate control put being ineffective until all the mixer
sources go to mute, we simply unhook the sources, set the rate and re-apply.

Change-Id: I31a97cccead0579b87e80b0be012b03905bd1776
Signed-off-by: Nariman Poushin <nariman@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>